Cleaning up physics demo and adding new scene
